src/WebApp/requirements.txt
Flask==0.12.1
# Flask-Sockets==0.2.1
Flask-Breadcrumbs==0.4.0
numpy==1.14.0
scipy==1.0.0
azure-storage==0.36.0
azure-batch==3.0.0
azure-common==1.1.8
azure-mgmt==1.0.0
pycryptodome==3.4.3
pyyaml==3.12
azure_servicebus==0.21.1
azure-iothub-device-client==1.3.2
azure-iothub-service-client==1.3.2
Markdown==2.6.11
keyring=11.0.0
